tests refactoring.
